+++ l18n/__init__.py
@@ -1,4 +1,4 @@
 from .version import __version__, __version_info__
 
-from .maps import tz_locations, territories
+from .maps import tz_cities, territories
 from .translation import set_language