+++ l18n/maps.py
@@ -2,8 +2,8 @@ from .translation import L18NDict
 
 try:
     from . import __maps
-    tz_locations = L18NDict(__maps.tz_locations)
+    tz_cities = L18NDict(__maps.tz_cities)
     territories = L18NDict(__maps.territories)
 except ImportError:
-    tz_locations = {}
+    tz_cities = {}
     territories = {}