+++ tests/translation.py
@@ -14,7 +14,7 @@ class TranslationTestCase(TestCase):
         l18n.set_language(None)
 
     def test_tz_city_translated(self):
-        self.assertEqual(str(l18n.tz_locations['Canada/Newfoundland']),
+        self.assertEqual(str(l18n.tz_cities['Canada/Newfoundland']),
                          u'Terre-Neuve')
 
     def test_territory_translated(self):