+++ CHANGES.rst
@@ -6,6 +6,6 @@ v2014.10.1 (dev)
 ----------------
 
 - Birth
-- Exposes ``tz_locations`` and ``territories`` dictionary-like objects
+- Exposes ``tz_cities`` and ``territories`` dictionary-like objects
 - Compatible with python 2.6+ and 3.3+
 - English and French translations